refactor: qualify test utility class
Renames and moves the QueryTest class since it's a general test helper.
