+++ qa/integration-tests/src/test/java/io/camunda/it/client/AdHocSubprocessActivitySearchTest.java
@@ -7,8 +7,8 @@
  */
 package io.camunda.it.client;
 
-import static io.camunda.it.client.QueryTest.deployResource;
-import static io.camunda.it.client.QueryTest.waitForProcessesToBeDeployed;
+import static io.camunda.it.util.TestHelper.deployResource;
+import static io.camunda.it.util.TestHelper.waitForProcessesToBeDeployed;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
