+++ qa/integration-tests/src/test/java/io/camunda/it/operate/OperatePermissionsIT.java
@@ -10,7 +10,7 @@ package io.camunda.it.operate;
 import static io.camunda.client.protocol.rest.PermissionTypeEnum.READ_PROCESS_DEFINITION;
 import static io.camunda.client.protocol.rest.PermissionTypeEnum.READ_PROCESS_INSTANCE;
 import static io.camunda.client.protocol.rest.ResourceTypeEnum.PROCESS_DEFINITION;
-import static io.camunda.it.client.QueryTest.deployResource;
+import static io.camunda.it.util.TestHelper.deployResource;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.client.CamundaClient;
