+++ qa/integration-tests/src/test/java/io/camunda/it/tasklist/v1/tenants/TasklistV1MultiTenancyIT.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.it.tasklist.v1.tenants;
 
-import static io.camunda.it.client.QueryTest.deployResourceForTenant;
+import static io.camunda.it.util.TestHelper.deployResourceForTenant;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.OK;
